requirements.txt
requests
pyreadline  # Only for Windows
pgpy
pycryptodome
colorama
standard-imghdr
